# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: parking API
  description: |
    API REST de gestión de plazas de parking de ALEATICA.
    Autenticación por cookie de sesión (`parking_SESSION`). Fase 1: login local.
    Fase 2: SSO ALEATICA (la landing autentica, parking autoriza).
  version: 1.0.0
  contact:
    name: Equipo parking
    email: "_[pendiente]_"
servers:
  - url: http://localhost:8080/parking-api/api/v1
    description: Desarrollo local
  - url: https://parking.aleatica.com/parking-api/api/v1
    description: Producción
...
security:
  - sessionCookie: []
paths:
...
  /employees/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    put:
      tags: [Employees]
      operationId: updateEmployee
      summary: Modifica un empleado (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdate'
      responses:
        '200':
          description: Empleado actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
    delete:
      tags: [Employees]
      operationId: deactivateEmployee
      summary: Baja lógica de un empleado (ADMIN)
      description: Marca `active=false`; no borra la fila.
      responses:
        '204': { description: Empleado dado de baja }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
...
  /fixed-assignments/employee/{employeeId}:
    parameters:
      - name: employeeId
        in: path
        required: true
        schema: { type: integer, format: int64, minimum: 1 }
    get:
      tags: [FixedAssignments]
      operationId: getEmployeeFixedAssignments
      summary: Asignaciones fijas de un empleado
      description: >
        ADMIN cualquiera; EMPLOYEE solo las suyas (verificación de objeto). Devuelve la
        lista plana de asignaciones activas de AMBOS tipos de recurso. Para el prefill
        del modal de empleado, el cliente agrupa por `resourceType`: las filas `PARKING`
        reconstruyen la plaza fija ({ parkingSpaceId, dayOfWeek[] }) y las `DESK` el
        puesto fijo ({ parkingSpaceId, dayOfWeek[] }). Un empleado puede tener plaza y
        puesto el mismo día (recursos independientes).
      responses:
        '200':
          description: Asignaciones fijas del empleado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FixedAssignment'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
    put:
      tags: [FixedAssignments]
      operationId: setEmployeeFixedAssignments
      summary: Asigna/modifica la asignación fija de un empleado (ADMIN)
      description: >
        Una plaza no puede asignarse a dos empleados el mismo día; un empleado no
        puede tener dos recursos el mismo día (409). Aplica solo desde la fecha
        actual; no afecta a días pasados ni a aprobaciones previas.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FixedAssignmentPutRequest'
      responses:
        '200':
          description: Asignación fija establecida
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FixedAssignment'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
    delete:
      tags: [FixedAssignments]
      operationId: revokeEmployeeFixedAssignment
      summary: Revoca (lógicamente) la asignación fija de un empleado (ADMIN)
      description: >
        Marca `active=false`, `revoked_at`, `revoked_by_id`. No borra. El parámetro
        opcional `resourceType` acota la revocación a un tipo de recurso: si se omite se
        revocan TODAS las asignaciones activas del empleado (comportamiento
        retrocompatible); si se indica (`PARKING` o `DESK`) se revocan SOLO las de ese
        tipo, dejando intactas las del otro. Devuelve 404 cuando no hay ninguna
        asignación activa del tipo solicitado.
      parameters:
        - name: resourceType
          in: query
          required: false
          description: >
            Tipo de recurso a revocar. Omitido = todos los tipos (default
            retrocompatible). `PARKING` revoca solo la plaza fija; `DESK` solo el puesto.
          schema:
            $ref: '#/components/schemas/ResourceType'
      responses:
        '204': { description: Asignación revocada }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }

  # ---------------------------------------------------------------------------
  # Releases
  # ---------------------------------------------------------------------------
...
  /push/subscriptions:
    post:
      tags: [Push]
      operationId: subscribePush
      summary: Registra la suscripción Web Push del usuario (autenticado)
      description: >
        Change `push-notifications`. Alta idempotente (upsert por `endpoint`) de la suscripción del
        navegador del usuario autenticado. Un usuario puede tener varias (multi-dispositivo).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushSubscriptionRequest'
      responses:
        '204': { description: Suscripción registrada }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
    delete:
      tags: [Push]
      operationId: unsubscribePush
      summary: Da de baja una suscripción Web Push del usuario (autenticado)
      description: >
        Change `push-notifications`. Baja por `endpoint` de una suscripción del propio usuario. No
        hace nada si el endpoint no existe o pertenece a otro empleado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushUnsubscribeRequest'
      responses:
        '204': { description: Suscripción dada de baja (o inexistente) }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
  /push/vapid-public-key:
    get:
      tags: [Push]
      operationId: getVapidPublicKey
      summary: Clave pública VAPID para suscribir el navegador (autenticado)
      description: >
        Change `push-notifications`. Devuelve la clave pública VAPID (base64url) que el cliente usa
        en `PushManager.subscribe`. La privada nunca se expone. Vacía si push no está configurado.
      responses:
        '200':
          description: Clave pública VAPID vigente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VapidKeyResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: parking_SESSION
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
...
  responses:
    BadRequest:
      description: Datos inválidos
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: Sin permisos para la operación
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: Conflicto de estado o unicidad
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ServerError:
      description: Error interno
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
...
    Role:
      type: string
      enum: [ADMIN, EMPLOYEE]
    AuthOrigin:
      type: string
      enum: [LOCAL, ENTRA_ID]
    EmployeeCategory:
      type: string
      description: >-
        Categoria jerarquica (rango organizativo) del empleado. Orden de mayor a
        menor rango: CEO, CONSEJO, DIRECTOR_N1, DIRECTOR_N2, GERENTE,
        MANDO_INTERMEDIO, EMPLEADO.
      enum: [CEO, CONSEJO, DIRECTOR_N1, DIRECTOR_N2, GERENTE, MANDO_INTERMEDIO, EMPLEADO]
...
    ApiError:
      type: object
      required: [error, message, timestamp]
      properties:
        error:
          type: string
          description: Código corto de error.
        message:
          type: string
        fields:
          type: object
          additionalProperties: { type: string }
          description: Errores por campo (validación).
        timestamp:
          type: string
          format: date-time

    # ---- Pagination ----
...
    Employee:
      type: object
      properties:
        id: { type: integer, format: int64 }
        firstName: { type: string }
        lastName: { type: string }
        login: { type: string }
        email: { type: string, format: email }
        department: { type: string, nullable: true }
        mobilePhone: { type: string, nullable: true }
        licensePlate: { type: string, nullable: true }
        isCorporate: { type: boolean }
        authOrigin: { $ref: '#/components/schemas/AuthOrigin' }
        role: { $ref: '#/components/schemas/Role' }
        category: { $ref: '#/components/schemas/EmployeeCategory' }
        enabled: { type: boolean }
        active: { type: boolean }
        passwordMustChange: { type: boolean }
        emailNotificationsEnabled:
          type: boolean
          description: 'Recibe avisos por email (change push-notifications); por defecto true.'
        pushNotificationsEnabled:
          type: boolean
          description: 'Recibe avisos por push (change push-notifications); por defecto true.'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time, nullable: true }
...
    EmployeeUpdate:
      type: object
      required: [category]
      properties:
        firstName: { type: string, minLength: 1, maxLength: 100 }
        lastName: { type: string, minLength: 1, maxLength: 150 }
        email: { type: string, format: email, maxLength: 255 }
        department: { type: string, maxLength: 100 }
        mobilePhone: { type: string, maxLength: 30 }
        licensePlate: { type: string, maxLength: 15 }
        isCorporate: { type: boolean }
        role: { $ref: '#/components/schemas/Role' }
        category: { $ref: '#/components/schemas/EmployeeCategory' }
...
    ResourceType:
      type: string
      enum: [PARKING, DESK]
      description: Discriminador del tipo de recurso reservable.
...
    FixedAssignment:
      type: object
      properties:
        id: { type: integer, format: int64 }
        parkingSpaceId: { type: integer, format: int64 }
        employeeId: { type: integer, format: int64 }
        dayOfWeek: { type: integer, minimum: 1, maximum: 7 }
        active: { type: boolean }
        createdById: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        revokedById: { type: integer, format: int64, nullable: true }
        revokedAt: { type: string, format: date-time, nullable: true }
        resourceType: { $ref: '#/components/schemas/ResourceType' }
    FixedAssignmentPutRequest:
      type: object
      required: [parkingSpaceId, daysOfWeek]
      properties:
        parkingSpaceId:
          type: integer
          format: int64
          description: 'Recurso a asignar: plaza (PARKING) o puesto (DESK).'
        daysOfWeek:
          type: array
          minItems: 1
          items: { type: integer, minimum: 1, maximum: 7 }
        resourceType:
          allOf: [ { $ref: '#/components/schemas/ResourceType' } ]
          default: PARKING
          description: Tipo de recurso; por defecto PARKING.
...
    PushSubscriptionRequest:
      type: object
      required: [endpoint, keys]
      description: >
        Suscripción Web Push del navegador (change `push-notifications`): el objeto que produce
        `PushManager.subscribe(...).toJSON()`.
      properties:
        endpoint:
          type: string
          description: Endpoint del push service (único).
          example: 'https://fcm.googleapis.com/fcm/send/abc123'
        keys:
          type: object
          required: [p256dh, auth]
          properties:
            p256dh:
              type: string
              description: Clave pública del cliente (base64url).
            auth:
              type: string
              description: Secreto de autenticación del cliente (base64url).
    PushUnsubscribeRequest:
      type: object
      required: [endpoint]
      description: 'Baja de una suscripción Web Push por endpoint (change `push-notifications`).'
      properties:
        endpoint:
          type: string
          description: Endpoint de la suscripción a dar de baja.
    VapidKeyResponse:
      type: object
      description: 'Clave pública VAPID (change `push-notifications`).'
      properties:
        publicKey:
          type: string
          description: Clave pública VAPID en base64url; vacía si push no está configurado.
          example: 'BEl62iUYgUiv...'
...